# server/src/interfaces/http/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a socco project",
    "version": "1.0.0",
    "title": "SOCCO Backend",
    "termsOfService": "",
    "contact": {
      "email": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
  "paths": {
...
    "/product/findBySearchText": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Finds Products by Search term",
        "description": "Return the list of products matching the search",
        "operationId": "findBySearchText",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "searchText",
            "in": "query",
            "description": "Status values that need to be considered for filter",
            "required": true,
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number of the result. Default value is 1.",
            "required": false,
            "type": "number",
            "collectionFormat": "multi"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "number of products in the result. Default value is 10.",
            "required": false,
            "type": "number",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
...
    "/community": {
      "get": {
        "tags": [
          "community"
        ],
        "summary": "Find all communities",
        "description": "Find all communities.",
        "operationId": "findCommunities",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "number of communities in the result. Default value is 10.",
            "required": false,
            "type": "number",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Community"
              }
            }
          },
          "400": {
            "description": "Communities not found"
          }
        }
      }
    },
    "/community/findBySearchText": {
      "get": {
        "tags": [
          "community"
        ],
        "summary": "Finds Community by Search term",
        "description": "Return the list of communities matching the search",
        "operationId": "findBySearchText",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "searchText",
            "in": "query",
            "description": "Status values that need to be considered for filter",
            "required": true,
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number of the result. Default value is 1.",
            "required": false,
            "type": "number",
            "collectionFormat": "multi"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "number of products in the result. Default value is 10.",
            "required": false,
            "type": "number",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Community"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "summary": "Get customer preferences",
        "description": "Get customer preferences by id",
        "operationId": "getPreferences",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Preferences"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          },
          "404": {
            "description": "Consumer id not found"
          }
        }
      },
      "put": {
        "tags": [
          "preferences"
        ],
        "summary": "Update customer preferences",
        "description": "Update customer preferences by id",
        "operationId": "updatePreferences",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "in": "body",
          "name": "body",
          "description": "Add shipping information for user",
          "required": true,
          "schema": {
            "$ref": "#/definitions/Preferences"
          }
        }],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Preferences"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          },
          "404": {
            "description": "Consumer id not found"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Products": {
      "type": "object",
      "required": [
        "products",
        "pageCount",
        "itemCount",
        "limit",
        "currentPage"
      ],
      "properties": {
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        },
        "pageCount": {
          "type": "number"
        },
        "itemCount": {
          "type": "number"
        },
        "limit": {
          "type": "number"
        },
        "currentPage": {
          "type": "number"
        }
      }
    },
    "Product": {
      "type": "object",
      "required": [
        "name",
        "description",
        "price",
        "discountedPrice",
        "display"
      ],
      "properties": {
        "productId": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "example": "Chartres Cathedral"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "discountedPrice": {
          "type": "number"
        },
        "image": {
          "type": "string"
        },
        "image2": {
          "type": "string"
        },
        "thumbnail": {
          "type": "string"
        },
        "display": {
          "type": "integer"
        }
      }
    },
...
    "Community": {
      "type": "object",
      "properties": {
        "agent_id": {
          "type": "integer",
          "format": "int64"
        },
        "community_name": {
          "type": "string"
        },
        "community_address": {
          "type": "string"
        },
        "agent_designation": {
          "type": "string"
        }
      }
    },
    "Preferences": {
      "type": "object",
      "properties": {
        "customer_id": {
          "type": "integer",
          "format": "int64"
        },
        "community_id": {
          "type": "integer",
          "format": "int64"
        },
        "community_name": {
          "type": "string"
        },
        "language_code": {
          "type": "string"
        }
      }
    }
  }
}